secured with protected method & added filtering.

+++ list/Item.java
@@ -2,10 +2,10 @@ package list;
 
 public class Item {
     private String item_name;
-    private double inventory_count;
+    private int inventory_count;
     private double value;
 
-    public Item(String item_name, double value, double inventory_count) {
+    public Item(String item_name, double value, int inventory_count) {
         this.item_name = item_name;
         this.value = value;
         this.inventory_count = inventory_count;
@@ -20,11 +20,11 @@ public class Item {
         return value;
     }
 
-    protected double get_Inventory() {
+    protected int get_Inventory() {
         return inventory_count;
     }
 
-    protected void set_Inventory(double count) {
+    protected void set_Inventory(int count) {
         inventory_count = count;
     }
